% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad .Renviron in case it contains an http_proxy which is not allowed at the IAE}
\usage{
.onLoad(libname, pkgname)
}
\description{
Auto-configure for students. Restarting R is mandatory when needed
}
