% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.fn.config}
\alias{IAE.fn.config}
\title{Configure python using list of field.}
\usage{
IAE.fn.config(course = "ALL.ALL", my.env = "")
}
\arguments{
\item{course}{An (optional) named argument to check package installation.}

\item{my.env}{An optional virtual environment for the python local installation.}
}
\value{
None
}
\description{
Configure python using list of field.
}
