% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.fn.pkg_install}
\alias{IAE.fn.pkg_install}
\title{Install a python package after verification.}
\usage{
IAE.fn.pkg_install(name = "", my.env = "")
}
\arguments{
\item{name}{The name of the package.}

\item{my.env}{An optional virtual environment for the python local installation.}
}
\value{
TRUE if the package was not already installed and FALSE if it was.
}
\description{
Install a python package after verification.
}
