% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M1.AAC}
\alias{IAE.M1.AAC}
\title{Configuration for course in Master 1 spécialité AAC}
\usage{
IAE.M1.AAC()
}
\description{
Auto-configure for students, installation of necessary packages are done as needed.
}
