% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M1.CCA}
\alias{IAE.M1.CCA}
\title{Configuration for course in Master 1 spécialité CCA}
\usage{
IAE.M1.CCA()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
