% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M1.MFI}
\alias{IAE.M1.MFI}
\title{Configuration for course in Master 1 spécialité Finance parcours Marchés Financiers}
\usage{
IAE.M1.MFI()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
