% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M1}
\alias{IAE.M1}
\title{Configuration for launching spyder only in the env}
\usage{
IAE.M1()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
