% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M1.teacher}
\alias{IAE.M1.teacher}
\title{Configuration for course in Master 1 for teacher}
\usage{
IAE.M1.teacher()
}
\description{
Auto-configure, installation of necessary packages are done as needed.
}
