% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.M2.MFI}
\alias{IAE.M2.MFI}
\title{Configuration for course in Master 2 Finance spécialités Marchés Financiers}
\usage{
IAE.M2.MFI()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
