% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.quarto}
\alias{IAE.quarto}
\title{Configuration without launching spyder for use in quarto}
\usage{
IAE.quarto()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
