% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.spyder}
\alias{IAE.spyder}
\title{Configuration foronly launching spyder in the env}
\usage{
IAE.spyder()
}
\description{
Auto-configure for students in their AppData, installation of necessary packages are done as needed.
}
