% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.teacher}
\alias{IAE.teacher}
\title{Configuration for course for teacher}
\usage{
IAE.teacher()
}
\description{
Auto-configure, installation of necessary packages are done as needed.
}
