% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad .Renviron in case it contains an http_proxy which is not allowed at the IAE}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{name of the lib}

\item{pkgname}{name of the package}
}
\value{
Message for restarting R when invalid proxy are found at the IAE.
}
\description{
.onLoad .Renviron in case it contains an http_proxy which is not allowed at the IAE
}
