% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.fn.config}
\alias{IAE.fn.config}
\title{Configure python using list of field.}
\usage{
IAE.fn.config(course = NULL, my.env = "")
}
\arguments{
\item{course}{An (optional) list of named arguments which are used to check package installation.}

\item{my.env}{An optional virtual environment for the python local installation.}
}
\value{
Currently none formaly but TRUE or FALSE depending on the installation of the spyder package which is mandatory here.
}
\description{
Configure python using list of field.
}
