% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localPythonUsingReticulate.R
\name{IAE.fn.env}
\alias{IAE.fn.env}
\title{Fix env from ~ expansion}
\usage{
IAE.fn.env()
}
\value{
Either a string corresponding to a subdirectories in AppData r-reticulate package directory or NULL when ~ cannot be extended for a user.
}
\description{
Fix env from ~ expansion
}
